event_manager/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.8.1
    # via
    #   -c C:\Users\user1\django_projects\event_project\event_manager\requirements.txt
    #   django
cfgv==3.4.0
    # via pre-commit
distlib==0.3.9
    # via virtualenv
django==5.2.3
    # via
    #   -c C:\Users\user1\django_projects\event_project\event_manager\requirements.txt
    #   django-debug-toolbar
django-debug-toolbar==5.2.0
    # via -r requirements-dev.in
factory-boy==3.3.3
    # via -r requirements-dev.in
faker==37.4.0
    # via factory-boy
filelock==3.18.0
    # via virtualenv
identify==2.6.12
    # via pre-commit
nodeenv==1.9.1
    # via pre-commit
platformdirs==4.3.8
    # via virtualenv
pre-commit==4.2.0
    # via -r requirements-dev.in
pyyaml==6.0.2
    # via pre-commit
ruff==0.12.1
    # via -r requirements-dev.in
sqlparse==0.5.3
    # via
    #   -c C:\Users\user1\django_projects\event_project\event_manager\requirements.txt
    #   django
    #   django-debug-toolbar
tzdata==2025.2
    # via
    #   -c C:\Users\user1\django_projects\event_project\event_manager\requirements.txt
    #   django
    #   faker
virtualenv==20.31.2
    # via pre-commit
